SDKJAVA-103: Rename common module to core
