+++ core/src/main/java/com/inrupt/client/core/IOUtils.java
@@ -18,7 +18,7 @@
  * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
  * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package com.inrupt.client.common;
+package com.inrupt.client.core;
 
 import java.io.IOException;
 import java.io.InputStream;
