+++ openid/src/main/java/com/inrupt/client/openid/OpenIdProvider.java
@@ -23,7 +23,7 @@ package com.inrupt.client.openid;
 import static java.nio.charset.StandardCharsets.UTF_8;
 
 import com.inrupt.client.authentication.DPoP;
-import com.inrupt.client.common.URIBuilder;
+import com.inrupt.client.core.URIBuilder;
 import com.inrupt.client.spi.JsonProcessor;
 import com.inrupt.client.spi.ServiceProvider;
