+++ pom.xml
@@ -31,7 +31,7 @@
   <modules>
     <module>authentication</module>
     <module>build-tools</module>
-    <module>common</module>
+    <module>core</module>
     <module>http</module>
     <module>jackson</module>
     <module>jena</module>
