+++ vc/src/main/java/com/inrupt/client/vc/Issuer.java
@@ -20,8 +20,8 @@
  */
 package com.inrupt.client.vc;
 
-import com.inrupt.client.common.IOUtils;
-import com.inrupt.client.common.URIBuilder;
+import com.inrupt.client.core.IOUtils;
+import com.inrupt.client.core.URIBuilder;
 import com.inrupt.client.spi.JsonProcessor;
 import com.inrupt.client.spi.ServiceProvider;
